Split utility functions into separate modules
